+++ include/pland/Global.h
@@ -5,9 +5,11 @@
 #include "mc/legacy/ActorUniqueID.h"
 #include "mc/platform/UUID.h"
 #include <atomic>
+#include <expected>
 #include <filesystem>
 #include <optional>
 #include <unordered_map>
+
 
 class Player;
 
@@ -55,8 +57,8 @@ LDNDAPI extern std::string GetPlayerLocaleCodeFromSettings(Player& player); // P
 inline int constexpr GlobalSubLandMaxNestedLevel = 16; // 子领地最大嵌套层数
 
 
-template <typename T>
-using Result = std::pair<T, std::optional<std::string>>; // T: 返回值, std::string: 错误信息
+template <typename T, typename E = std::string>
+using Result = std::expected<T, E>;
 
 } // namespace land